requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
flask-sqlalchemy==3.0.3
SQLAlchemy==2.0.7
idna==3.4
gunicorn==20.1.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.3
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3